+++ lib/ui/onboarding2/Onboarding2Widgets.dart
@@ -1,6 +1,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:illinois/service/Analytics.dart';
+import 'package:illinois/ui/widgets/FocusableSemanticsWidget.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/triangle_painter.dart';
@@ -91,8 +92,11 @@ class Onboarding2ToggleButton extends StatelessWidget{
 
   @override
   Widget build(BuildContext context) {
-    return
-      Semantics(
+    return FocusableSemanticsWidget(onSelect: () {
+          onTap!();
+          anaunceChange();
+        },
+        child: Semantics(
           label: _label,
           value: (toggled!
               ? Localization().getStringEx(
@@ -123,7 +127,7 @@ class Onboarding2ToggleButton extends StatelessWidget{
                   ),
                 )
             ),],),
-          ));
+          )));
   }
 
   void anaunceChange() {