+++ lib/ui/widgets/RoleGridButton.dart
@@ -17,6 +17,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter/semantics.dart';
+import 'package:illinois/ui/widgets/FocusableSemanticsWidget.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:illinois/service/FlexUI.dart';
 import 'package:rokwire_plugin/service/localization.dart';
@@ -84,18 +85,18 @@ class RoleGridButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) =>
-    GestureDetector(onTap: _onTap, child:
-      Semantics(label: title, excludeSemantics: true, sortKey: _semanticsSortKey, value: _semanticsValue, child:
-        selected ? Stack(children: [
-          _contentWidget,
-          Positioned.fill(child:
-            Align(alignment: Alignment.topRight, child:
-              _selectionMarker
+    FocusableSemanticsWidget(onSelect: _onTap, child: GestureDetector(onTap: _onTap, child:
+        Semantics(label: title, excludeSemantics: true, sortKey: _semanticsSortKey, value: _semanticsValue, child:
+          selected ? Stack(children: [
+            _contentWidget,
+            Positioned.fill(child:
+              Align(alignment: Alignment.topRight, child:
+                _selectionMarker
+              )
             )
-          )
-        ],) : _contentWidget
-     )
-    );
+          ],) : _contentWidget
+        )
+    ));
 
   Widget get _contentWidget => Padding(padding: margin, child:
     Padding(padding: margin, child: